adds utilities for vectors of 64-bit reals
NOTE:
These were created by replacing the types and kinds in the argument
lists and subroutine names.

This is a staged commit at a point where the vector class has been
tested for compilation and execution. The tests were done to check
for minimal functionality, and these tests will be committed later.

Note that the implementation source files for vectors containing
64-bit floating-point numbers (reals) will be committed separately.