removes code that tests erase method [vector]
COMMENTS:
Removes code used for testing the `experimental' erase method of
the vector class.

A backup of the source code that includes all tests have been
added to the repository for future use.

Again, this makes it easier to push to the repo the improvements
that have been made without the `experimental' method.