improves makefile organization [vector]
COMMENTS:
Partitions `Vector-inc' into other include files, one for the
methods `Vector-methods-inc', another for the implementations
`Vector-implements-inc', and the last one `make-inc' which
includes all the others.

The makefile has been setup to include `make-inc'. This has been
done so that the makefiles that build codes for testing require
minimal revision. All that was done on them was to include the
mentioned includes of the vector class.

The code compiles and executes successfully.