refactors the push-back method [vector] <int32_t>
COMMENTS:
Refactors similar lines of code into the generic `push' method
that pushes a value unto the back of the vector. Note that the
method will not be truly generic until applying the respective
changes to vectors that contain objects of different types.

Introduces the global parameter `VECTOR_MIN_SIZE' for convenience.
Later on one can easily set the minimum to another value without
having to edit many lines of code.

Makes the `create' subroutine eaiser to read.