partitions vector copy method [vector]
COMMENTS:
Copying data from vector into array has been replaced by the
`backup' method, which does exactly what's needed. So the lines
of code used for copying the data into the placeholder arrays have
been deleted.

The only (minor) difference is that the `backup' method copies the
data in vector delimited by the asymmetric range [0, `avail');
whereas the original source used the symmetric range 0:`avail'. For
this reason the assignment to the upper bound `ub' has been modified
accordingly.

Note that `avail' is a valid off-the-end index but it holds whatever
that happens to be in memory when the internal array is allocated.

Also since `backup' handles memory allocation of the placeholder
arrays the `ary_bounds' variable is no longer needed so it has
been also deleted from the source.