partitions the assignment operator [vector]
COMMENTS:

Extends the push-back method for arrays and uses it for
partitioning further the assignment operator. Note that the method
assumes the vector has enough space for the array.

As usual the code is written manually for vectors containing 32-bit
integers and `sed' commands are used to produce the code for other
vector types.

Minor additional modifications have been made to the assignment
operator. The internal data of the (destination) vector is destroyed
before `vector_vector_t_copy' which handles copying the contents
of the source into the destination vector.

The internal procedure `alloc' has been renamed to `clone' because
that's the most important action that it does. It ``clones'' the
``identity'' of the internal data of the source vector.

The internal procedure `push' has also been renamed to `copy' for
disambiguation since it shares the name of the generic `push'
method used to push the arrays into the vectors. Even though there
are generics sharing the name `copy' these are not actually used
and might be removed in the future.

The internal procedure `bind' has been removed because the generic
`push' method already binds the iterator to the internal array.