frees (destination) vector on assignment [vector]
COMMENTS:

This revision has been introduced in response to issue #84

Assigning a vector to another means that the existing data on
the (destination) vector needs to be destroyed first.

To this end the ``allocators'' of each of the vector components
(iter, data, etc.) first check for their allocation status. If
it turns out that memory has been allocated, the ``allocators''
free the memory via a call to `deallocate'. Note that the
`destructor' handles the memory deallocation of the components
of the types (iter, data, etc.) so that the ``allocator'' does
no overt work. (This is why we do not explicitly deallocate the
components elsewhere but in the destructor.) Also note that
the iterator is nullified by its `destructor' as it should be.

Thus, whenever the ``instantiate'' subroutine is invoked, we know
that any existing data is going to be destroyed because it forwards
the memory management to the ``allocators''. This is how we make
sure the (destination) vector is destroyed upon assignment because
the assignment operator invokes ``instantiate'' early on.

The code compiles, executes, and passes the memory leak tests.