implements `fill' constructor [vector]
COMMENTS:
Implements a constructor that creates a `vector<int32_t>' filled
with `n' copies of a (given) value.

The subroutine `vector_int32_t_push_n_copies' has been adapted
from the `vector_int32_t_push_array` subroutine. Instead of
calculating the number of elements `numel', we set `numel' to be
equal to the number of copies. And we just copy the value `value'
into all the respective elements of the vector.

The internal state of the vector is set to `true' because the vector
contains at least a value. The default constructor does not set the
variable because it creates an empty vector. That's the difference
between these constructors.

Note that if we don't set the internal state, a subsequent call to
the push-back method will destroy the data. Memory allocated for
the internal array will be freed and the vector will be resized
to its minimum size by the `create' subroutine.