adds `fill' constructor for vectors of [vectors]
COMMENTS:
As usual, ``sed'' commands were used to generate the code from
a suitable `template'. I am not reporting them since they are
very similar to others which have been documented in the commit
messages.

In contrast to the implementation of the `fill' constructors for
vectors of built-in types (integers, reals, etc.) the constructor
needs to validate the iterators of the contained vectors. (The
object is a vector of vectors so its elements are also vectors).

It turns out that it's not enough to simply recreate the iterators
of each one of the elements (vectors) in the `fill' constructor,
it is necessary to do so in the subroutine that implements the
``assignment'' operator.

To confirm this I used the method ``addr'' to print the addresses
of both the iterator and the internal array of the vectors. (Note
that this method has been implemented for the first time to debug
the code.) That's what led me to uncover the problem:

the iterators were pointing to the same memory address.

One could say that it's not necessary to validate the iterators
in the `fill' constructor because these ``outlive'' the data they
point to. And they would be right because in the end the vector
returned by the `fill' constructor is assigned to another vector
and destroyed.

Nevertheless, validating the iterators in the `fill' constructor
helped troubleshoot the issue so I am committing the work done for
completeness.

The root of the problem is solved by validating the iterators in
the `vector_vector_t_copy' subroutine which implements the
``assignment'' of a vector into another.

Note that the validation of the iterators is not carried out
recursively so that the validation step is expected to work only
for vectors whose elements are vectors of built-in types. I need
to keep working on it and have an idea how to do it. I was kind
of in a hurry and wanted a quick fix so that's the reason for
committing.

This is, nevertheless, a step towards fixing issue #62

Anyone interested in what tests were used to troubleshoot the code
may check the changes done to the `test_vector.for' source file.