validates iterators recursively [vector]
COMMENTS:
Re-associates iterators recursively to validate them so that they
point to the (actual) internal array of the vector.

Points out in the test code `test_vector.for' that the iterators
point to an invalid object --- an object that has already been
deallocated from memory.

This commit has been used to investigate further issue #62 and I
am referencing it here for documentation purposes.