sets error message in `fill' constructor [vector]
COMMENTS:
This change has been made in response to issue #87

Without it the user won't receive a useful diagnostic message
about what went wrong when attempting to push mixed-types into
a vector of built-in types.

Memory is allocated for the internal error message in the subroutine
`vector_vector_t_copy' which implements the ``assignment'' operator,
as it should have been done.