caters nested `empty' [vectors] when validating
COMMENTS:
Accounts for `empty' vectors nested in another vector when
validating the iterators of a vector of vectors.

This is of importance when a vector is ``assigned'' to another
vector and when using the ``fill'' constructor to create a vector
of `empty' vectors; note that the vector class already has ways
to deal with `empty' vectors in other situations.

Note that an `empty' vector may have its components allocated or
not so the `validation' subroutine has to cater the possibility of
a vector whose components have no memory allocated for them; and
this is handled by calling the subroutine ``instantiate'' (see
line 318 of `Vector_type_oblivious_methods.for').

Moreover, instead of complaining and aborting execution if an
`empty' vector is found, the code simply nullifies its iterator
(line 360 of the same source file as above). Note that catering
vectors of mixed types is always done before the validation step,
so it's fine to use ``class default'' to match `empty' vectors.

Note that an `empty' vector is a valid vector nonetheless whose
internal array has yet to be `molded' to a type supported by the
vector class. At a later time the user can push a supported value
to the `empty' vector. And the ability to do just that might come
in handy later on ...

The changes have been made so that the test code won't crash when
trying to associate to a unallocated internal array (line 323 of
`Vector_type_oblivious_methods.for'). The crash would be otherwise
triggered on line 176 of `test_vector.for' where a vector of
`empty' vectors is created.