caters for `empty' vectors (revised) [vectors]
COMMENTS:
This revision has been made in response to issue #90

It turns out that the previous commit still needed revision since
it was not working as expected --- the iterators of the `empty'
vectors were pointing to invalid memory locations (out-of-bounds).

I am refering to line 336 of `Vector_type_oblivious_methods.for'.
At that line the bounds `b' and `e' would be equal to `0' and `-1',
respectively, so the iterator would end up pointing to an invalid
memory address instead of being nullified. (I should have tested
the code further before pushing it to the repository.)

Additionally, note that the internal array of an `empty' vector is
not allocated even after a call to the ``instantiate'' subroutine.
Thus, associating to an unallocated component is not something
we want to do (line 323 of the same source file).

The usage of ``class default'' in line 360 has been reverted to
detect unexpected errors because it shouldn't match (and it does
not match) `empty' vector anymore.

The changes made to the code `fix' the mentioned issues brought by
the previous commit. This was verified by the testing code in
lines 175 -- 212 of `test_vector.for' (this commit).

For testing, I could have checked for the association status of the
iterators but opted to check for the memory addresses, which for a
nullified pointer in FORTRAN is zero.

I have also kept around the code used for debugging, it might be
of some use later. It has been defined as an internal procedure
named `debug'. Note that it's only meant for vector of vectors.

I did not indent some lines of code so that the `git diff' would
be more informative of the actual changes.

The mentioned line numbers refer to the lines of the source files
as they were in the previous commit unless otherwise specified.