extends the push-back method for arrays <int32_t>
COMMENTS:
A step towards completing request #63

Used the existing framework for pushing a 32-bit integer to extend
the push-back method [vector] for arrays of the same type.

The array size is taken into account when computing the limits
of the vector. To this end, the subroutine `double_vector_size'
accepts an optional input argument that specifies how much space
is needed to accomodate the array. It uses that info instead of the
present limit to expand the vector.

Testing shows that the vector grows as needed and that it stores
the elements of the arrays pushed unto its back.

The code has also been tested for memory leaks with `valgrind'. No
issues have been detected thus far.