implements range constructor for vector <int32_t>
COMMENTS:
The range constructor creates a vector from an asymmetric range,
for example [0, n) where `0' and `n' are the `begin' and `end'
values, respectively. Note that the `end' value is excluded from
the sequence in contrast to the symmetric range [0, n].
The number of elements of an asymmetric range is determined by the
difference of the `end' and `begin' values.

The constructor takes one positional argument which corresponds to
the `end' value. The `begin' and `step' are optional arguments just
like Python's range. If the user does not provide values for the
optional arguments, the constructor sets `begin' to zero and `step'
to one (just like Python's range).

The constructor has been partitioned into internal procedures. Of
particular interest the `counter' procedure caters invalid input
arguments, as for example, the step is a positive integer but the
sequence is decreasing. (Without proper handling the code would
loop forever.) If the step is zero then it is set to one to avoid
division by zero and the `end' and `begin' are made equal to one
another so that we obtain an `empty' range.

Regarding `empty' ranges, the constructor still allocates memory
so that the vector has the minimum capacity for storage, and, as
one would expect, it does not fill the internal array with any
values and the iterator points to NULL. The ``validate'' method
which validates the iterator has been adapted to account for
`empty' vectors with preallocated memory.

The number of elements that the range has is determined by the
difference of the `end' and `begin' values divided by the `step'.
Note that the remainder of the difference:

		mod(`end' - `begin', `step')

is extracted to obtain the actual `last' value where the range
would `end'. Since we know exactly where the range would `end'
we can use an inequality to control how many elements are pushed
unto the vector (line 170). (The construction of the while loop
is analogous to what is done in c++ to iterate over the elements
of a vector.)

We do not use `numel' to update the `avail' iterator, instead we
use a counter to determine how many elements have been pushed.
Note that `numel' may not represent the number of elements if the
range has fewer elements than the minimum storage size of the
vector or if the range is empty. Thus the vector has either the
minimum size or a size that matches the number of elements in the
range. Contrary to other constructors of the class which allocate
at most twice the size needed. The reason for doing this is that
you may just want a quick way to generate a range but have no
intention of including more values. (At least that's how I intend
to use it.) So in a way the vector is tailored to `fit' unless the
number of elements is less than the minimum vector capacity.

The constructor delegates the task of associating the iterator to
any data or nullifying whatever the case may be to the ``validate''
method.

Finally the constructor sets the state of the vector to signal that
the vector has been molded to a type and that it's ready to store
more values.