implements range constructor for vector <int64_t>
COMMENTS:
Used similar `sed' commands to generate great portions of the code
to add support for vectors <int64_t>.

Had to define an asymmetric-range class to be able to `overload'
the vector constructor to create vectors <int64_t> from a range.

It's still possible to create a vector <int32_t> as in the previous
commit. I wanted to keep that way of doing that for its simplicity.

A couple of `minor' issues were fixed:

defining a module procedure as a `plain' subroutine in a submodule

The Intel FORTRAN Compiler detected those instances.

The code passes all tests and valgrind did not report memory leaks.