adds support for vector<pointer_t> [vector]
COMMENTS:
A vector <pointer_t> can be interpreted as a dynamic array of
pointers. It could be used as the underlying type for a random
access iterator of a linked-list. Note that the iterator of a
linked-list in C++ is either forward or bidirectional. How would
the random-access iterator would perform is something that needs
to be checked though.

Most of the code was generated using `sed' commands similar to
those used in prior commits.

WARNING:
Intel FORTRAN Compiler ifort (IFORT) 2021.1 Beta 20201112
fails to produce a binary that passes the tests for vectors of
pointers vector<pointer_t>. (I'm using a Beta so it's possible
that it might have been fixed.)

On the other hand, GNU FORTRAN Compiler 10.3.0 produces a binary
that passes all the tests.

The problem appears to be how the Intel FORTRAN Compiler treats a
pointer to a class(*) object. A pointer shold point to the memory
address of an object, and that's what any programmer would expect.
Instead of that, ifort appears to be copying the data into another
class(*) object. On line 1011 of `test_vector.for' I am using the
`loc' intrinsic to print the memory address of the pointer and the
class(*) object and to my surprise I found them to be different.