adds support for removing the last element [iter]
COMMENTS:
Adds code to remove the last element from a vector / iterator.

Caters self-assignment for the pointer type and implements a
destructor as well.

As a consequence of that some of the tests in [vector] had to
be adapted accordingly to avoid memory leaks (dangling pointers).
Valgrind reports no instances after the change.