adds random-acesss iterator & append method [list]
COMMENTS:
Implements a random-access iterator for the linked-list class.
And the append method has been implemented as well. These
are to be regarded as initial implementations.

The code passes test successfully: the data pointed to by the
iterator is consistent with the inserted values.

Valgrind reports no memory leaks.

Comments out tests that won't be conducted for the time being.