caters empty links< > [list]
COMMENTS:
If the link is NULL, it makes sense to nullify the pointer to the
node as it is done. If there's a node pointed to by the link, then
the node pointer is associated to the linked node (as before). The
change accounts for the possibility of an empty link< >.