implements valide-iterator method [list]
COMMENTS:
Validating the iterator is needed if an element is either added
or removed from the list.

The initial implementation uses recursive procedures but these can
overflow the stack for long lists so the implementation that
traverses the list (via forward iterators) has been selected
because is cheaper in terms of memory usage.

Nevertheless, both versions are committed to the repository; just
bear in mind that only the non-recursive one is used by the list.

By following the same train of thought, the link ``destructor'' is
no longer a recursive procedure. The downside is that it is
considerably slower than its recursive counterpart because it only
destroys the `tail' of the list on every iteration. So one has to
traverse the list as many times as the number of elements in the
list.

All the versions of the destructor are present in this commit.