tests validating the list iterator [list]
COMMENTS:
The test consist of appending more values to the list followed by
iterator validation.

The test shows that the iterator is properly validated. The values
pointed to by the iterator are consistent with the values used to
generate the list.

Valgrind reports no memory leak issues.

It might take a while to destroy the list ... especially when using
valgrind.

Some minor changes like changing the print format have been made.