adds interface to use a C linked-list from FORTRAN
COMMENTS:
The C linked-list is a forward linked-list.

The FORTRAN forward linked-list class simply encapsulates the
C linked-list. Its methods forward their work to the lower-level C
functions.

The code tests the append method of the FORTRAN forward linked-list
class. Even though the class is missing an iterator to get at the
values, there is code that shows how to gets at the values of the
C forward linked-list from FORTRAN. That code will be adapted in
a future commit to provide an iterator.

Valgrind reports no memory leaks.