improves forward linked-list [FORTRAN]
COMMENTS:
The O(N*N) algorithm for destroying nodes has been replaced with
a O(N) algorithm.

It also removes the [recursive] destructors since these won't be
used.

It's worth mentioning that the C FORTRAN interoperable linked-list
destroys the nodes about ten times faster than this implementation.

Note that it's the responsibility of the node destructor to
deallocate the data contained in the unlimited polymorphic pointer
of the `pointer_t'. The destructor of the `pointer_t' type has
been modified so that it only nullifies the pointer, it won't
free allocated resources.