adds support for appending 64-bit integers [list]
COMMENTS:
Applies to the forward linked-list written in C.

Adds support for inserting signed 64-bit integers at the end of
the linked-list.

Caters for inserting mixed-type integers to linked-list. The code
aborts execution gracefully without incurring on memory leaks if
a mixed-type is input to the append method. A diagnostic message
is written to the standard error stream on such instances.

Linked-list passes the existing tests.