a more direct approach for generating the iterator
COMMENTS:
Users of the class have the option of creating their own method for
obtaining a random-access iterator for the linked-list while
avoiding the overhead of creating a generic iterator as was done
in the previous commits. They are in a better position to do so
because they know the data type contained in the linked-list.

One could implement a method that does the same thing on behalf of
the user by considering all the types supported by the linked-list
(possibly via a switch construct) ... but that is a task for later.