destroys iterator automatically [list] FORTRAN
COMMENTS:
The destructor of the FORTRAN forward linked-list class delegates
the task of destroying the iterator to the lower-level C function
if it's associated. Thus, users of the class need not to worry
about freeing the resources allocated for the iterator.

The users can still destroy it manually if they need to though.
Note that appending new data into the list invalidates the iterator
so a new one might need to be generated to be able to get at the
values contained in the list.