compares array to list runtime performance [list]
COMMENTS:
It requires less time to create an array having the same values as
the list including the time required for memory allocation as one
might expect. And destroying the array takes even less time.

This is just a simple test evincing the expected outcome. More
interesting tests shall be conducted later on.