uses system clock for profiling [list] C
COMMENTS:
The CPU times are comparable to those reported by the FORTRAN
forward linked-list which is built upon these lower-level C
functions. Therefore, we can deduce that the FORTRAN API incurs
a negligible overhead.

It would be interesting to see the output generated by profiling
tools (such as gprof) to conduct a more thorough assessment of the
runtime performance.

RUNTIME:
Implemented an append method that only creates a node at the end
without the sane checks (not committed) and it turns out that the
runtimes are similar. This means that the CPU time spent on doing
the checks can also be safely neglected and that memory allocation
of the nodes might be the bottleneck (10x slower than allocating
an array of built-in type).