discards ``dead weight'' linked-list [list] C
COMMENTS:
The `data_t' simply made the linked-list less compact, more
complicated to use, and slower, partly because of the need to
allocate memory for the `data_t' object and also because of
possible CPU cache misses.

RUNTIME:
Creating the random-access iterator is almost as fast as creating
an array of built-in type having as many elements as the iterator.
Generating the linked-list is about 20x slower than creating an
array containing the same values, and destroying the list is about
50x slower than freeing an array with the same number of elements.
Bjarne Stroustrup's ``stay compact'' advice says it all.

Note that the runtimes are wild guesses based on a couple of runs.
Moreover, pitting a linked-list against an array on a ``sprint'' is
not a fair competition. What needs to be compared is the algorithm.

For example, comparing a particle sorting algorithm that uses a
linked-list against another that uses an array would be interesting
to conduct as a function of the system size and particle number
density. There might be conditions where a list-based algorithm
outperform its array-based counterpart, or even better, support the
array-based algorithm.