adds absolute position vector [clang] [FORTRAN]
COMMENTS:
adds absolute position vector

The absolute position vector is independent of the periodic
boundary conditions, that is, it is equal to the position vector
that the particle would have in the absence of periodicity (as if
it were unbounded).

This simplifies computation of the MSD for periodic systems. This
shall be added in a future commit.

code passes tests

valgrind reports no memory issues