adds initial neighbor-list code [clang] [FORTRAN]
COMMENTS:
The BDS code passes all existing and new tests.

On purpose all particles are in contact with one another to test
the implementation of the algorithm that generates the neighbor
list.

This version does not take into account the periodicity of the
system; nevertheless, it is correct to use it for dense suspensions
where all the particles interact with one another (or form a single
cluster sort of speak). Of course, the list won't be of any help
other than to check that indeed all particles interact with one
another.

Periodicity considerations shall be added later.

Valgrind reports no memory issues.

MACROS:

CONTACT is the contact distance which is two for unit spheres.

RANGE is the interaction range of the particles, and we use its
squared value to avoid having to compute the squared root in the
code for speed.

SANE CHECKS:

We have limited the maximum number of particles due to the use
of negative numbers to store the ID of the head (or root) of the
neighbor-list. Note that in the code we add `one' to the `id' so
that if `id' is already the maximum (signed) 64-bit integer the
operation would overflow, thereby breaking the logic. Nevertheless,
I do not think that anyone would attempt a simulation of that
scale in a serial code.