makes list() more concise [clang]
COMMENTS:
Note that link() already caters when two nodes are already linked,
so it is no different if the same node is passed as argument to
the method.

The time complexity still cubic, O(N^3), where N is the number
of particles. Note that search for the root node runs on linear
time.

BDS code passes the existing tests