initial list() impl passes new test [FORTRAN]
COMMENTS:
demonstrates that when there is more than one cluster of particles
some of the particles do not interact with one another, as one
might expect.

the test consisted on exporting the particle positions to a plain
text file when there is more than one cluster in the system and
then checking (later) that there are indeed particles that do
not interact with other particles.

have yet to consider the periodicity of the system

have added some convenient MACROS

BDS code passes new and existing tests

valgrind reports no memory issues