fixes misleading FAIL test outcome msgs [FORTRAN]
COMMENTS:
After consideration of the periodic boundaries, the tail of the
neighbor-list (or linked-list array) is no longer the last element.
What we do now is that there should be only one cluster of linked
particles because all the particles are at contact in a grid-like
or (lattice) structure.

Since the last element of the linked-list array is no longer the
tail we have to traverse all the elements of the list to conduct
the search (without excluding the last which holds one of them).