relocates time-related MACROS
COMMENTS:
Just to get rid of some implicit conversions and warnings issued
by the GNU FORTRAN Compiler.

Want to make sure that there's no loss of precision in the time
step (there's no more a 32 to 64-bit implicit conversion).

In the future it might be useful to define the number of steps
as a 64-bit integer for long simulations.