adds shifted Lennard-Jones SLJ force [clang]
COMMENTS:
The force is zero if the distance is zero and if the distance is
greater than the interaction range. This is so that we can use
square for-loops to compute the net force on the particles. This
is easier to parallelize with POSIX threads and OpenMP and it is
also easier to vectorize than the triangular counterpart (yet with
the same quadratic time complexity).

Note that only the repulsive contribution of the Lennard-Jones
potential is used to compute the force. The force is shifted so
that it vanishes exactly at the radius of cutoff (RANGE).

valgrind reports no memory issues