fixes SLJ time splitting [clang]
COMMENTS:
Makes the Shifted Lennar-Jones SLJ force better suited for conducting
Brownian Dynamics simulations; that is, the characteristic time
of particle displacement due to the SLJ force is not too different
from that comes from the Stochastic forces.

One could say that in the former SLJ force caused a time splitting
that would make BDS longer than necessary so that the code properly
captures the behavior of the particles.

code passes existing tests