adds headers as dependencies [Makefile]
COMMENTS:
Changes to `system.h' means that the whole project needs to be built,
for many source files depend directly on it and those that don't
depend on it directly have dependencies that do so in the end the
whole project needs to be recompiled.

Formerly a change to `system.h' had no effect and so it was necessary
to cleanup the project via `make clean' and then issue `make' to
rebuild the project. This is clearly not the right thing to do but
since I am the only developer and there are no users (yet) it did
not bother me to work this way ...

This commit fixes that issue.