updates PRNG seeder() [clang]
COMMENTS:
for the time being XORs the current time and the process ID to seed
the PRNG

complains if the seed is inadequate (that is, equal to zero) or if
the time utility fails unexpectedly

complains at compile time if the underlying type of `time_t' is not
64-bits wide

uses bitmasking so that the caller can check if the seeding failed;
if the Least Significan Bit LSB is set (with a non-zero value) that
means that seeding failed and that the code should abort execution.

Note that Marsaglia's xorshift pseudo-random number generators must
never be seeded with a zero value (for you will always fetch a zero
value from the PRNG).

code passes existing tests