adds runtime checks [FORTRAN]
COMMENTS:
Note that the types in question are C types whose definitions can
be found in the header file particles/sphere/type.h.

If these types are actually modified later on, one has to update
the (FORTRAN) API accordingly.

The runtime checks act like reminders to the next programmer.

The static asserts in the C counterpart of this test shall be
helpful in producing an error message if these types are edited.