adds the OBDS sphere_t [OOP] [FORTRAN]
COMMENTS:
The OBDS sphere_t defines the deferred (or abstract) methods of the
abstract OBDS particle_t. These methods forward their work to their
respective implementations in C.

The OBDS sphere_t constructor returns a pointer. The use of an
allocatable otherwise is inconvenient because as soon as the
OBDS sphere_t object (in the constructor) goes out of scope the
destructor obliterates everything ... and the code would segault.
When using a pointer we are responsible for freeing the memory
allocated for the object via a call to deallocate, and this is
exactly what we want. (The destructor gets called behind scenes.)

Note that we do not use the assignment operator but associate to
the pointer returned by the constructor instead.

The BDS code has been tested with GCC 10.3.0 and 13.2.0

valgrind reports no memory issues