adds sphere properties [clang] [FORTRAN]
COMMENTS:
adds the displacement vector, the director (or orientation vector),
the neighbor-list, and some placeholders for storing intermediate
computations.

uses bitfields to pad the OBDS_Sphere_t so that it is aligned to
a 64-byte boundary (could have padded enough only for a 16-byte or
a 32-byte alignment but opted for a 64-byte alignment).

adds assertion (at compile-time) that checks that the code is
being compiled in a system that stores data in a little endian
fashion, for the code already assumes that's the case (according
to GLIBC's bits/endianness.h all x86_64 CPUs are little-endian).

the displacement vector and the neighbor-list shall be used later
to improve the performance of the OBDS code.

the FORTRAN MODULE API has been updated accordingly.