logs the resultant force [clang]
COMMENTS:
the resultant encompasses the stochastic and deterministic
forces acting on the particles

for the time being we use memcpy() to store the deterministic
forces in the `neighbor-list', we have made sure that the neighbor
list has enough space to store the forces

the space complexity of the OBDS application is now O(N log N) owing
to the space alloted for the neighbor-list

later we shall use the neighbor-list to actually store the list
of neighbors, but for now we are content with using it as a
placeholder for the deterministic forces

instead of defining the number of particles in a header, what we
do is to store the power of two (2^8 = 256), in FORTRAN we use the
built-in exponentiation operator ** to obtain the number of spheres
and in C we use the left shift operator << to do the same. This is
done this way so that we can perform some static assertions to
catch errors early on.

code passes existing tests

valgrind reports no memory issues