adds time-averaged MSD computation [Python]
COMMENTS:
Uses the exported spheres data in contrast to the initial BDS code
which did the computation on the fly

At the present the Python code computes the rotational Mean Squared
Displacement MSD but it can be easily extended to study the linear
counterpart.

The computation of the time-averaged MSD might take hours ... all
because there's a lot of IO involved.

In practice, one does not use the results from a single BDS run but
several. The Python code only uses the data from a single run. Despite
of this there's a nice agreement between the theory and the MSD results
from a single run.

Just bear in mind that there are as many elements in the MSD array
as data files if you are interested in looking at the implementation.