writes particle data to tempfiles [clang]
COMMENTS:
Writes particle data to a tempfile and if all goes well (no IO ERRORS)
then the file is renamed.

We don't want to process the data of a BDS run if there are temp files
lingering around.

Another use for this is self-checkpointing. If we attempt to restore
a simulation from an incomplete data file (due to some interruption
of sorts) the code would fail. But if instead, we use temporaries
we know that (it is highly unlikely) for the data in the txt file
to be incomplete (or corrupted). This is because the IO is done
on the tmp file and we rename it only after we are done with it.

We don't need to close the file descriptor to the temporary file
because fclose() already does that for us.