updates endianness check [clang]
COMMENTS:
In MAC OS X's endiah.h `__BYTE_ORDER` and `__LITTLE_ENDIAN` are
not defined with double underscores as shown but instead as
`BYTE_ORDER` and `LITTLE_ENDIAN`

This commit takes that into account. OK so when we added code to
compile the source with the Intel compilers we broke it by using
the double underscored MACROS ... this is not an elegant solution
and we hate to admit it but it works and that's what matters.
This is the first time we are attempting to write portable code.

Things like this can always be improved down the road.

DEV NOTES:
We are using an if-else conditional because we do not want to skip
the endianness assertion. We start by assuming that we are compiling
in GNU/Linux (and that is highly likely because that's what the
HPC runs) and we also know that if the OS is GNU/Linux then `endian.h`
is provided by GLIBC and so the double underscore MACROS must be
defined. These are not defined then we might be compiling in OSX
or another POSIX compliant OS. If the assertion fails we might be
compiling the source in a non POSIX compliant OS so we also want
to know that at compile time.